fix: use getBrief in fm test
